pcapng, tcpdump: enable deferral and reconnect
If you specify deferred connect = true, we'll spin off a
thread rather than making the fifo open() fail if there is
no reader.

If you specify reconnect = true, we'll drop the packet dumps
 when the reader goes away, but re-establish it and resume
dumping if/when the reader comes back.  Tearing down the
gate hook (or changing or removing the FIFO path, so that
open() produces an error rather than waiting for a reader)
will make the hook stop capturing packets.